Allow Model.define to be passed module
This will auto destroy the model on unload and derive the model name
from module.id

+++ app/bart/model/server-main.js
@@ -26,6 +26,7 @@ define(function(require, exports, module) {
     },
 
     destroyModel: function (model, drop) {
+      if (! model) return;
       if (drop === 'drop')
         mongoDb.defaultDb.dropCollection(model.modelName);
       model.docs = null;
